+++ types/domain/entities/Issue.d.ts
@@ -25,5 +25,6 @@ export type Issue = {
     createdAt: Date;
     author: string;
     closingIssueReferenceUrls: string[];
+    agent: string | null;
 };
 //# sourceMappingURL=Issue.d.ts.map