+++ types/domain/usecases/adapter-interfaces/ProjectRepository.d.ts
@@ -11,6 +11,9 @@ export interface ProjectRepository {
     updateStatusList: (project: Project, statusOptions: (Omit<FieldOption, 'id'> & {
         id: FieldOption['id'] | null;
     })[]) => Promise<FieldOption[]>;
+    updateAgentList: (project: Project, agentOptions: (Omit<FieldOption, 'id'> & {
+        id: FieldOption['id'] | null;
+    })[]) => Promise<FieldOption[]>;
     getByUrl: (url: string) => Promise<Project>;
 }
 //# sourceMappingURL=ProjectRepository.d.ts.map